requirements.txt
aiogram>=3.0.0
gspread>=5.10.0
python-dotenv>=1.0.0
apscheduler>=3.10.4